requirements.txt
streamlit
requests
beautifulsoup4
pytest
reportlab
python-docx
